chore(directory structure): refactor and modify directories structure
