+++ package.json
@@ -9,7 +9,7 @@
     "open:src": "babel-node tools/srcServer.js",
     "lint": "node_modules/.bin/esw webpack.config.* client tools",
     "lint:watch": "npm run lint -- --watch",
-    "test": "istanbul cover _mocha ./test/testSetup.js \"./test/**/*.spec.js\"",
+    "test": "istanbul cover _mocha ./server/test/testSetup.js \"./server/test/**/*.spec.js\"",
     "test:watch": "npm run test -- --watch",
     "db:migrate": "node_modules/.bin/sequelize db:migrate"
   },
