+++ server/app.js
@@ -3,9 +3,6 @@
 import express from 'express';
 import parser from 'body-parser';
 import homeRoute from './routes/index';
-import documentRoutes from './routes/document';
-import userRoutes from './routes/user';
-import roleRoutes from './routes/role';
 
 require('dotenv').config();
 
@@ -22,8 +19,5 @@ app.use(parser.urlencoded({ extended: false }));
 
 // Specifying the routes the app should use for different purposes
 app.use('/', homeRoute);
-app.use('/role', roleRoutes);
-app.use('/document', documentRoutes);
-app.use('/user', userRoutes);
 
 export default app;
