+++ server/test/helper.js
@@ -1,3 +1,5 @@
+/* eslint-disable import/no-extraneous-dependencies */
+
 import faker from 'faker';
 
 module.exports = {
