+++ server/test/models/role.spec.js
@@ -2,7 +2,7 @@
 
 import { expect } from 'chai';
 import helper from '../helper';
-import model from '../../server/models';
+import model from '../../models';
 
 const Role = model.Role;
