+++ server/test/server/role.spec.js
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
-import app from '../../server';
-import models from '../../server/models';
+import app from '../../app';
+import models from '../../models';
 import helper from '../helper';
 
 const request = require('supertest')(app);
